requirements.txt
aniso8601==6.0.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
blinker==1.4
Click==7.0
dateutils==0.6.6
Flask==1.0.2
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-mongoengine==0.9.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.7
Flask-Security==3.0.0
Flask-WTF==0.14.2
importlib-metadata==0.18
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mongoengine==0.17.0
mongoengine-goodjson==1.1.6
more-itertools==7.1.0
packaging==19.0
passlib==1.7.1
Pillow==6.0.0
pluggy==0.12.0
py==1.8.0
pymongo==3.7.2
pyparsing==2.4.0
pytest==5.0.1
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
speaklater==1.3
Unipath==1.1
wcwidth==0.1.7
Werkzeug==0.15.2
WTForms==2.2.1
xlrd==1.2.0
zipp==0.5.2
